Vulkan_1_3_Barriers (#90)
* brdf lut

* activating synchronization2

* culling

* aabb

* blit

* cleanup

* wip

* memory and buffer barriers

* image barrier

* VulkanBarrier.cpp

* removing warning messages

+++ HelloVulkan/Source/Scene.cpp
@@ -47,7 +47,7 @@ void Scene::CreateBindlessResources(VulkanContext& ctx)
 	{
 		for (size_t j = 0; j < models_[i].meshes_.size(); ++j)
 		{
-			meshDataArray_.emplace_back(models_[i].meshes_[j].GetMeshData(textureIndexOffset, i));
+			meshDataArray_.emplace_back(models_[i].meshes_[j].GetMeshData(textureIndexOffset, static_cast<uint32_t>(i)));
 		}
 		textureIndexOffset += models_[i].GetNumTextures();
 	}
